# src/test/resources/Collection_Service_API_Spec_v1.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "https://<host>:<port>/cp/collection-service/",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/v1/node-profiles": {
            "get": {
                "tags": [
                    "node-profile-controller"
                ],
                "summary": "Get all Node Profile name and id for landing page",
                "operationId": "getAllNodeProfile",
                "responses": {
                    "200": {
                        "description": "Found the Node Profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AllNodeProfileData"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error getting the node profile details."
                    }
                }
            },
            "put": {
                "tags": [
                    "node-profile-controller"
                ],
                "summary": "Update the NodeProfile along with the node list and node filter.",
                "operationId": "updateNodeProfile",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NodeProfileData"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Updated the Node Profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NodeProfileData"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Node Profile not found"
                    },
                    "500": {
                        "description": "Error updating the Node Profile"
                    }
                }
            },
            "post": {
                "tags": [
                    "node-profile-controller"
                ],
                "summary": "Create a new node profile. Node List and node filter is optional at the time of creation.",
                "operationId": "createNodeProfile",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NodeProfileData"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Node profile is created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NodeProfileData"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error creating the node profile."
                    }
                }
            },
            "delete": {
                "tags": [
                    "node-profile-controller"
                ],
                "summary": "Delete a Node Profile by its ID. The associated node lists and node filter are also deleted.",
                "operationId": "deleteNodeProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Node Profile is deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NodeProfileData"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Node Profile not found"
                    },
                    "500": {
                        "description": "Error deleting the node profile"
                    }
                }
            }
        },
...
        "/api/v1/node-profiles/info-and-node-filter": {
            "put": {
                "tags": [
                    "node-profile-controller"
                ],
                "summary": "Update the NodeProfile along with node filter only. The node list is not updated.",
                "operationId": "updateNodeProfileInfoAndNodeFilter",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NodeProfileData"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Updated the Node Profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NodeProfileData"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Node Profile not found"
                    },
                    "500": {
                        "description": "Error updating the Node Profile"
                    }
                }
            }
        },
...
        "/api/v1/node-profiles/{id}": {
            "get": {
                "tags": [
                    "node-profile-controller"
                ],
                "summary": "Get the given NodeProfile details including all its node lists and node filter by its ID",
                "operationId": "getNodeProfileById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found the Node Profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NodeProfileData"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Node Profile not found"
                    },
                    "500": {
                        "description": "Error getting the node profile."
                    }
                }
            }
        },
        "/api/v1/node-profiles/info": {
            "get": {
                "tags": [
                    "node-profile-controller"
                ],
                "summary": "Get the basic NodeProfile details  by its ID. This does not include the node lists and node filter associated with node profile.",
                "operationId": "getNodeProfileInfoById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found the Node Profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NodeProfileData"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Node Profile not found"
                    },
                    "500": {
                        "description": "Error getting the node profile info."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NodeFilterData": {
                "required": [
                    "condition",
                    "type",
                    "value"
                ],
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "HOST_INDIVIDUAL",
                            "HOST_REGEX",
                            "IP_CIDR",
                            "IP_REGEX",
                            "IP_INDIVIDUAL"
                        ]
                    },
                    "value": {
                        "type": "string"
                    },
                    "condition": {
                        "type": "string",
                        "enum": [
                            "INCLUDE",
                            "EXCLUDE"
                        ]
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "NodeListData": {
                "required": [
                    "nodeIp"
                ],
                "type": "object",
                "properties": {
                    "nodeIp": {
                        "type": "string"
                    },
                    "nodeManagementIp": {
                        "type": "string"
                    },
                    "authGroupName": {
                        "type": "string"
                    },
                    "snmpGroupName": {
                        "type": "string"
                    },
                    "snmpGroupMismatch": {
                        "type": "boolean",
                        "readOnly": true
                    },
                    "authGroupMismatch": {
                        "type": "boolean",
                        "readOnly": true
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "NodeProfileData": {
                "required": [
                    "name"
                ],
                "type": "object",
                "properties": {
                    "updateDate": {
                        "type": "string",
                        "format": "date",
                        "readOnly": true
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "defaultAuthGroup": {
                        "type": "string"
                    },
                    "defaultSnmpGroup": {
                        "type": "string"
                    },
                    "useNodeListAsIncludeFilter": {
                        "type": "boolean"
                    },
                    "nodeList": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NodeListData"
                        }
                    },
                    "nodeFilter": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NodeFilterData"
                        }
                    }
                }
            },
...
            "AllNodeProfileData": {
                "type": "object",
                "properties": {
                    "updateDate": {
                        "type": "string",
                        "format": "date",
                        "readOnly": true
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "useNodeListAsIncludeFilter": {
                        "type": "boolean"
                    }
                }
            },
...
        }
    }
}